+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Added
 * worker: Add grep-able state transitions to the log ("Entering STARTED/READY/UNHEALTHY state")
 * bigip-fast-templates: Large drop-down lists can be filtered by searching with associated text field
+* worker: Allow installing template sets from GitHub
 
 ## Fixed
 * bigip-fast-templates: Correctly reference existing snatpool in Exchange template (issue [#106](https://github.com/F5Networks/f5-appsvcs-templates/issues/106))